+++ test/should_index_test.rb
@@ -39,7 +39,6 @@ class ShouldIndexTest < Minitest::Test
     end
     Product.where(id: product.id).reindex
     Product.searchkick_index.refresh
-    # TODO fix in Searchkick 5
-    # assert_search "index", []
+    assert_search "index", []
   end
 end